Documentation Updates
On the header, add links to the Solution Homepage and GitHub Repository.

In the contents side-pane, add a new section for Additional
Documentation and a link to the Implementation Guide to make it more
discoverable.

On the welcome page:
* Add sections explaining the split of documentation between the
implementation guide/git-managed docs, and a pointer to getting started

Update wording in the "Data Model" page from Perspective -> Workload
Discovery.

Remove the Cost Optimization page, as this is still in the
Implementation Guide so is duplicated.